# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "CETPRO MDLM API",
        "description": "API REST para el Sistema de Gestión del Centro de Educación Técnico Productiva María de Lourdes Muñoz.\n\n## Módulos Disponibles\n- **Estudiantes**: Gestión de estudiantes y apoderados\n- **Matrículas**: Inscripciones, cronogramas de pago y pagos\n- **Académico**: Programas, cursos, especialidades y horarios\n- **RRHH**: Docentes y empleados\n- **Sistema**: Usuarios, roles y permisos\n\n## Autenticación\nEsta API no requiere autenticación para fines de desarrollo. En producción, se recomienda implementar autenticación via Sanctum o Passport.",
        "contact": {
            "name": "Soporte CETPRO",
            "email": "soporte@cetpro-mdlm.edu.pe"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://cetpro-mdlm.test/api/v1",
            "description": "Servidor Laravel Herd"
        }
    ],
    "paths": {
...
        "/pagos/{id}": {
            "get": {
                "tags": [
                    "Pagos"
                ],
                "summary": "Obtener detalle de un pago",
                "description": "Retorna la información detallada de un pago específico incluyendo datos del cronograma, matrícula y estudiante.",
                "operationId": "31c15295bca2631483c5c9c15e881ac3",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del pago",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Pago encontrado exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "monto": {
                                                    "type": "number",
                                                    "format": "float",
                                                    "example": 150
                                                },
                                                "fecha_vencimiento": {
                                                    "type": "string",
                                                    "format": "date",
                                                    "example": "2024-12-31"
                                                },
                                                "fecha_pago": {
                                                    "type": "string",
                                                    "format": "date",
                                                    "example": "2024-12-15",
                                                    "nullable": true
                                                },
                                                "estado": {
                                                    "type": "string",
                                                    "enum": [
                                                        "pendiente",
                                                        "pagado",
                                                        "vencido",
                                                        "anulado"
                                                    ],
                                                    "example": "pagado"
                                                },
                                                "metodo_pago": {
                                                    "type": "string",
                                                    "example": "efectivo",
                                                    "nullable": true
                                                },
                                                "numero_cuota": {
                                                    "type": "integer",
                                                    "example": 1
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Pago no encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/NotFoundResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/pagos/{id}/registrar": {
            "post": {
                "tags": [
                    "Pagos"
                ],
                "summary": "Registrar pago de una cuota",
                "description": "Marca una cuota como pagada registrando el método de pago y opcionalmente la evidencia del pago.",
                "operationId": "248e073ace8593d54c4adc05d9bca6e7",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del pago a registrar",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "description": "Datos del pago",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "metodo_pago"
                                ],
                                "properties": {
                                    "metodo_pago": {
                                        "description": "Método de pago utilizado",
                                        "type": "string",
                                        "enum": [
                                            "efectivo",
                                            "transferencia",
                                            "yape",
                                            "plin",
                                            "deposito"
                                        ],
                                        "example": "efectivo"
                                    },
                                    "evidencia_path": {
                                        "description": "Ruta del archivo de evidencia del pago (voucher, captura, etc.)",
                                        "type": "string",
                                        "example": "pagos/voucher_001.pdf",
                                        "nullable": true
                                    },
                                    "usuario_id": {
                                        "description": "ID del usuario que registra el pago (opcional, por defecto el usuario autenticado)",
                                        "type": "integer",
                                        "example": 1,
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Pago registrado exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "properties": {
                                                "message": {
                                                    "type": "string",
                                                    "example": "Pago registrado exitosamente"
                                                },
                                                "pago": {
                                                    "type": "object"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Pago no encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/NotFoundResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Error de validación",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/pagos/{id}/anular": {
            "post": {
                "tags": [
                    "Pagos"
                ],
                "summary": "Anular un pago",
                "description": "Marca un pago como anulado, revirtiendo su estado a pendiente.",
                "operationId": "58e8bfa1e3ae029abc46dc8735d45d39",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del pago a anular",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Pago anulado exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "properties": {
                                                "message": {
                                                    "type": "string",
                                                    "example": "Pago anulado exitosamente"
                                                },
                                                "pago": {
                                                    "type": "object"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Pago no encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/NotFoundResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Error al anular el pago",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/pagos/{id}/revertir": {
            "post": {
                "tags": [
                    "Pagos"
                ],
                "summary": "Revertir un pago",
                "description": "Revierte completamente un pago registrado, devolviendo el estado a pendiente. Requiere permisos especiales y un motivo justificado.",
                "operationId": "4da1f8da07df15cc0f866cbf42a9c4b1",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del pago a revertir",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "description": "Motivo de la reversión",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "motivo"
                                ],
                                "properties": {
                                    "motivo": {
                                        "description": "Motivo por el cual se revierte el pago",
                                        "type": "string",
                                        "example": "Error en el monto registrado"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Pago revertido exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "properties": {
                                                "message": {
                                                    "type": "string",
                                                    "example": "Pago revertido exitosamente"
                                                },
                                                "pago": {
                                                    "type": "object"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Pago no encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/NotFoundResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Error al revertir el pago",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
...
            "ErrorResponse": {
                "properties": {
                    "message": {
                        "type": "string",
                        "example": "Error al procesar la solicitud"
                    },
                    "errors": {
                        "description": "Detalles de los errores de validación",
                        "type": "object"
                    }
                },
                "type": "object"
            },
            "NotFoundResponse": {
                "properties": {
                    "message": {
                        "type": "string",
                        "example": "Recurso no encontrado"
                    }
                },
                "type": "object"
            }
        }
    },
...
}